+++ Cell2FireC/Forest.h
@@ -15,25 +15,26 @@
 using namespace std;
 
 // Class definition
-class Forest {
-    public:
-        int id;
-        int nCells;
-        double area;
-        double vol;
-        double age;
-        double perimeter;
-		std::string location;
-        std::vector<int> coord;	
-        std::unordered_map<std::string, int> fTypes;
-        std::vector<int> availCells;
-        std::vector<int> burntCells;
-        
-		// Constructor & methods
-        Forest(int _id, std::string _location, std::vector<int> _coord, int _nCells, double _area, 
-		          double _vol, double _age, double _perimeter, std::unordered_map<std::string, int> & _fTypes);
-        
-		void print_info();
+class Forest
+{
+public:
+    int id;
+    int nCells;
+    float area;
+    double vol;
+    double age;
+    float perimeter;
+    std::string location;
+    std::vector<int> coord;
+    std::unordered_map<std::string, int> fTypes;
+    std::vector<int> availCells;
+    std::vector<int> burntCells;
+
+    // Constructor & methods
+    Forest(int _id, std::string _location, std::vector<int> _coord, int _nCells, float _area,
+           double _vol, double _age, float _perimeter, std::unordered_map<std::string, int> &_fTypes);
+
+    void print_info();
 };
 
 #endif
